Move learn slash command to its own directory Augment learn response with optional Azure Resource Graph query result
